requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
firebase==4.0.1
firebase-admin==6.2.0
tensorflow==2.12.0
Flask==2.3.2
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0